Define rand defaults for AbstractProbabilisticProgram (#79)
This PR adds a 3-arg form of `rand` (suggested by @devmotion in https://github.com/TuringLang/DynamicPPL.jl/issues/466#issuecomment-1436670214) to the interface for `AbstractProbabilisticProgram` and implements the default 1- and 2-arg methods that dispatch to this.

Currently tests fail because this breaks the fallbacks for `GraphPPL.Model`, which expects `rand` to forward to its `rand!` method. I'm not certain how we want to define the interface for this `Model`.

Co-authored-by: Xianda Sun <sunxdt@gmail.com>